+++ src/main/java/agh/ics/oop/IMapElement.java
@@ -4,7 +4,7 @@ public interface IMapElement {
     public Vector2d getPosition();
     public String getUrl();
 
-    int getEnergy();
+    float getEnergy();
 
     String getEnergyUrl();
 }
